+++ package.json
@@ -7,7 +7,15 @@
     "AMQP 0-9-1",
     "RabbitMQ"
   ],
-  "main": "./channel_api.js",
+  "exports": {
+    ".": {
+      "default": "./dist/esm/index.js",
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js"
+    }
+  },
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
   "type": "module",
   "scripts": {
     "build": "node build.js",
