+++ src/index.js
@@ -1,1 +1,14 @@
-/// exportar modulos
+export { IllegalOperationError } from './error.js'
+import { IllegalOperationError } from './error.js'
+
+export * as credentials from './credentials.js'
+import * as credentials from './credentials.js'
+
+export { connect } from './api.js'
+import { connect } from './api.js'
+
+export default {
+  connect,
+  credentials,
+  IllegalOperationError
+}
